pyproject.toml
[project]
name = "warrior-and-demon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-xai>=1.1.0",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
